+++ packages/isaacscript-common/src/sets/charactersWithBlackHeartFromEternalHeartSet.ts
@@ -2,6 +2,6 @@ import { PlayerType } from "isaac-typescript-definitions";
 
 export const CHARACTERS_WITH_BLACK_HEART_FROM_ETERNAL_HEART_SET: ReadonlySet<PlayerType> =
   new Set([
-    PlayerType.BLACK_JUDAS, // 12
+    PlayerType.DARK_JUDAS, // 12
     PlayerType.JUDAS_B, // 24
   ]);