+++ packages/docs/docs/main/change-log.md
@@ -12,8 +12,12 @@ This page lists the changes to the IsaacScript framework.
 ## September 1st, 2022 (Unreleased)
 
 - `isaacscript init` now has a `--dev` flag for setting up a mod that will be testing out a development version of `isaacscript-common`. For more information, see the [README](https://github.com/IsaacScript/isaacscript#working-with-isaacscript-common).
+- Added the following helper functions:
+  - `removeAllNonAlphanumericCharacters`
 - Added the following custom commands:
   - `hush` (to warp to the Blue Womb Boss Room)
+  - `spawnCollectible` (to spawn a collectible by name)
+  - `spawnTrinket` (to spawn a collectible by name)
 
 ## August 25th, 2022